Split tests for easier maintenance
